requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
buildozer==1.5.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docutils==0.21.2
email_validator==2.2.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.6.0
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
Js2Py==0.74
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.gstreamer==0.3.4
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
lxml==6.0.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
multidict==6.4.3
oauthlib==3.3.1
openai==1.75.0
packaging==24.2
pexpect==4.9.0
pillow==10.4.0
pipwin==0.5.2
platformdirs==4.3.8
propcache==0.3.1
ptyprocess==0.7.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.11.3
pydantic_core==2.33.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.10.1
PyMsgBox==1.0.9
PyPDF2==3.0.1
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pyttsx3==2.98
pytweening==1.2.0
pywhatkit==5.4
pywin32==310
requests==2.32.3
requests-oauthlib==2.0.0
sh==2.2.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
tqdm==4.67.1
twilio==9.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
URLObject==2.4.3
virtualenv==20.31.2
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
WTForms==3.2.1
yarl==1.20.0
